app/requirements.txt
dj-database-url==0.5.0
Django==2.2.1
gunicorn==19.9.0
psycopg2
whitenoise==4.1.2
django-extensions==2.1.7
requests==2.22.0
django-debug-toolbar
django-debug-panel
pycpfcnpj==1.5.1
django-maintenance-mode==0.13.1
django-celery-email==2.0.1
Werkzeug

# Async
celery==4.3.0
django-celery-beat==1.5.0
redis==3.2.1
django-redis==4.5.0

# Channels
channels==2.2.0
channels-redis==2.4.0
daphne==2.3.0

# Compressor
django-compressor==2.2
django-compressor-toolkit==0.6.0

# Custom app settings
django-appconf==1.0.3
six

# REST
djangorestframework==3.9.4
django-filter==2.1.0
Markdown==3.1.1
djangorestframework-csv==2.1.0
djangorestframework-msgpack==1.0.2
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
rest-condition==1.0.3

# Social
facebook-sdk==3.1.0

# Storages
django-storages==1.7.1
boto3==1.9.112

# Monitoring and error handling
scout-apm==2.0.4
sentry-sdk==0.8.0
django-prometheus==1.0.15

# Analytics
django-analytical==2.5.0

# Progressive Web App
django-progressive-web-app==0.1.1

# Payments
six
python-dateutil
xmltodict
